Fix broken HTMLParser Code
Previously, the Code that parsed strings and turned them into React Elements
used the .innerHTML method to set the source code of a node. This broke in
certain situations because the "xmldom" package was used instead of a native
browser implementation. This commit fixes the issue, by directly passing the
html string to the document parser during creation.