Upgrade Dockerfile
This commit makes three changes to the Dockerfile.

This commit updates the node version used in the Dockerfile from 12 to
14. This is to bring it in line with the tests that are being written.

Next, it disables the yarn cache inside the docker container. This is to
decrease installation size.

Finally, it limits the user runnning the server to "www-data". This is
done to improve security in case.